PP-9498 Split AgreementNotFoundException into two
Split AgreementNotFoundException into
AgreementNotFoundException and
AgreementNotFoundBadRequestException.

The former is used with an exception mapper that maps it to a 404
(not found) response and is intended for requests to endpoints like
/v1/api/accounts/account-id/agreements/non-existent-agreement-id
/cancel. The latter maps to a 400 (bad request) and is intended for
requests to URLs like /v1/api/accounts/account-id/charges with a
JSON body containing "agreement_id": "non-existent-agreement-id".
