+++ src/main/java/uk/gov/pay/connector/agreement/exception/AgreementNotFoundException.java
@@ -1,8 +1,10 @@
-package uk.gov.pay.connector.charge.exception;
+package uk.gov.pay.connector.agreement.exception;
 import javax.ws.rs.WebApplicationException;
 
 public class AgreementNotFoundException extends WebApplicationException {
+
     public AgreementNotFoundException(String message) {
         super(message);
     }
+
 }
