+++ src/main/java/uk/gov/pay/connector/agreement/exception/AgreementNotFoundExceptionMapper.java
@@ -1,7 +1,8 @@
-package uk.gov.pay.connector.charge.exception;
+package uk.gov.pay.connector.agreement.exception;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import uk.gov.pay.connector.charge.exception.AgreementNotFoundBadRequestException;
 import uk.gov.pay.connector.common.model.api.ErrorResponse;
 import uk.gov.service.payments.commons.model.ErrorIdentifier;
 
@@ -17,13 +18,14 @@ public class AgreementNotFoundExceptionMapper implements ExceptionMapper<Agreeme
     @Override
     public Response toResponse(AgreementNotFoundException exception) {
         LOGGER.info(exception.getMessage());
-        
+
         ErrorResponse errorResponse = new ErrorResponse(ErrorIdentifier.AGREEMENT_NOT_FOUND, exception.getMessage());
 
-        return Response.status(Response.Status.BAD_REQUEST)
+        return Response.status(Response.Status.NOT_FOUND)
                 .entity(errorResponse)
                 .type(APPLICATION_JSON)
                 .build();
     }
+
 }
