+++ src/main/java/uk/gov/pay/connector/charge/exception/AgreementNotFoundBadRequestExceptionMapper.java
@@ -1,0 +1,30 @@
+package uk.gov.pay.connector.charge.exception;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import uk.gov.pay.connector.common.model.api.ErrorResponse;
+import uk.gov.service.payments.commons.model.ErrorIdentifier;
+
+import javax.ws.rs.core.Response;
+import javax.ws.rs.ext.ExceptionMapper;
+
+import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
+
+public class AgreementNotFoundBadRequestExceptionMapper implements ExceptionMapper<AgreementNotFoundBadRequestException> {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(AgreementNotFoundBadRequestExceptionMapper.class);
+
+    @Override
+    public Response toResponse(AgreementNotFoundBadRequestException exception) {
+        LOGGER.info(exception.getMessage());
+        
+        ErrorResponse errorResponse = new ErrorResponse(ErrorIdentifier.AGREEMENT_NOT_FOUND, exception.getMessage());
+
+        return Response.status(Response.Status.BAD_REQUEST)
+                .entity(errorResponse)
+                .type(APPLICATION_JSON)
+                .build();
+    }
+
+}
+
